+++ src/pyunrealsdk/hooks.h
@@ -12,6 +12,13 @@ namespace pyunrealsdk::hooks {
  * @param mod The module to register within.
  */
 void register_module(py::module_& mod);
+
+/**
+ * @brief Checks if we should automatically call inject next call for bound function calls.
+ *
+ * @return True if we should auto inject python calls.
+ */
+bool should_auto_inject_py_calls(void);
 #endif
 
 /**